paranoid check to ensure we're deleting what we think we're deleting

+++ file.go
@@ -58,7 +58,7 @@ func NewWithSuffix(filename, suffix string) (*File, error) {
 
 	file, err := os.OpenFile(filename, os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0644)
 	if err != nil {
-		cleanupFile(tempFile)
+		cleanupFile(tempFile, tempFilename)
 		return nil, err
 	}
 
@@ -97,7 +97,7 @@ func (w *File) Accept() error {
 		return fmt.Errorf("Content mismatch.  Expected OID %s, got %s", w.Oid, sig)
 	}
 
-	if err := cleanupFile(w.file); err != nil {
+	if err := cleanupFile(w.file, w.filename); err != nil {
 		w.Close()
 		return err
 	}
@@ -120,14 +120,14 @@ func (w *File) Accept() error {
 // Close cleans up the internal file objects.
 func (w *File) Close() error {
 	if w.tempFile != nil {
-		if err := cleanupFile(w.tempFile); err != nil {
+		if err := cleanupFile(w.tempFile, w.tempFilename); err != nil {
 			return err
 		}
 		w.tempFile = nil
 	}
 
 	if w.file != nil {
-		if err := cleanupFile(w.file); err != nil {
+		if err := cleanupFile(w.file, w.filename); err != nil {
 			return err
 		}
 		w.file = nil
@@ -144,7 +144,11 @@ func (w *File) Closed() bool {
 	return false
 }
 
-func cleanupFile(f fileWriteSyncer) error {
+func cleanupFile(f fileWriteSyncer, name string) error {
+	if fname := f.Name(); name != fname {
+		return fmt.Errorf("Invalid filename, expected %q, got %q", name, fname)
+	}
+
 	err := f.Close()
 	if err := os.RemoveAll(f.Name()); err != nil {
 		return err
